// types/360LeadsTypes.ts
export type leadsType = {
    id?: string;
    bmBranchCode: string;
    bmDomainId: string;
    clientBusinessAddress: string;
    clientCellCode: string;
    clientCellNumber: string;
    clientEmployerBusinessName: string;
    clientName: string;
    creatorId: string;
    creator_name: string;
};